Make injected meta tags configurable and change defaults
This change proposes 2 things:

1. Make apple-mobile-web-app-capable and apple-mobile-web-app-status-bar-style configurable
2. Change the defaults to "no" and "default", respectively

The reasoning for these changes have to do with the fact that those are nonstandard meta tags and many people do not fully understand what they are opting in to by setting "apple-mobile-web-app-capable" to "no". [This article](https://medium.com/@firt/dont-use-ios-web-app-meta-tag-irresponsibly-in-your-progressive-web-apps-85d70f4438cb) does a great job of explaining it, but essentially iOS versions before 11.3 use those tags to determine if they should be in "standalone" mode. But iOS < 11.3 has no back button, service worker, or even the ability to store history state when "apple-mobile-web-app-capable" is "yes". It is generally a good idea to just not do it.

Now that iOS 11.3 is properly supporting progressive web apps, it makes sense to only support PWAs in iOS 11.3+. All previous versions should opt out of Apple's nonstandard standalone mode. I think the default should be opt-out because it's the best practice, but the final decision should be left up to the user, thus why I want to make it configurable.