Change references to 'prefetch' plugin.
I have edited the code examples to reference the plugin by the name 'prefetch-app', as this seems to be the correct name the webpack config requires from inspecting the generated config. I have tested and can confirm this works with the example that deletes the plugin altogether, but not the blacklisting example.