Explain when .env files are loaded. (#4171)
It appears that .env files are only loaded at the start of `vue-cli-service` and are not watched for hot reloads.
**I am actually not sure this is the case for everyone, but is so for me.**

I think it is worth noting because the expectation is that a file at the root of the project would be watched for hot reloads.

(cherry picked from commit b22bedbf941dbf3d38ca6806e06758643bb1a768)