fix: do not generate empty file
The recent refactoring in #4330 introduced a regression where a `Home.vue` file was generated if using the TS plugin even without the router plugin. This happened because the result from the ejs rendering was `undefined`. This fixes the issue by returning an empty string instead, which will not be written.