fix: should take all configFiles and lockfiles into account
The previous implementation is based on the assumption that config files
have precendences, e.g. `.eslintrc.js` will take higher precendence
over `.eslintrc`, and only one will take effect.

This is not accurate however. For example, babel relies on both babel
config and browserslist config, so we need to deal with 2 config files.