fix: fix "Vue packages version mismatch" error caused by other global packages
closes #5161

The root cause of such issues:

`vue-jscodeshift-adapter` requires `vue-template-compiler`.
`vue-template-compiler` tries to require the `vue` package and check
its version on startup.

Normally, there's no `vue` package in the dependency tree of `@vue/cli`,
so this check will just skip.

But if the user has installed some other global package that depend on
`vue`, and hoists it to the root `node_modules`, `vue-template-compiler`
would successfully require it and check against that `vue` version,
which sometimes may be outdated, thus failing the bootstrap process.