fix(modern-mode): should use intersection of modern browsers and user-specified targets
We don't need to target browsers that users don't need, even if it is
a browser that supports `script type="module"`.

With this change, with the default target of "last 2 versions", we no
longer has to deal with the obsolete versions of evergreen browsers that
comes with all kinds of bugs.
This further improves the modern build bundle size. Maybe even better
than [the `bugfixes` option of `preset-env`](https://babeljs.io/docs/en/babel-preset-env#bugfixes)