Added link to S3 deploy cleanup plugin
I wrote a companion to the S3 deployment plugin that helps to cleanup the S3 bucket after deploying by tagging all files in the deployment directory in S3 that are not on the local machine. Thus, any files with old hashes will be tagged and can be expired in a set number of days using S3 lifecycle rules.

This helps prevent the bucket from becoming filled with old files, making it difficult to tell which is the most current version of a given file.

I added a link to this new plugin to the deployment docs -- please let me know your thoughts!