Correct nginx.conf root path
The documentation at ### Docker (Nginx) 4. Create a `nginx.conf` file in the root of your project it is suggested to use /app. 
It should be /usr/share/nginx/html;

This can be very frustrating for first time devs trying to run vue in docker as this would result in a 500 error page.