fix(cypress): make cypress a peer dependency
Cypress has a very stable CLI API.
Every 12 weeks Cypress releases a new Major version with
breaking changes to the JS tests and config.
But almost no changes to the CLI API.
This rhythm of Major updates does not match the slower pace of
the vue CLI.

To allow more nimble updates by end users, we add
cypress as a true devDependency to the user package.json
and as a peerDependency to send a warning if we remove it.