feat: add a @vue/cli-plugin-webpack-4 package for future use
More and more webpack plugins no longer support webpack 4 in their most
recent versions. For better webpack 5 compatibility and easier upgrade
in the future, we should update these dependencies.

So it makes sense to extract webpack-4-specific dependencies and logic
to a standalone package.

While the task is not that urgent at the moment, and I haven't got the
time to complete this refactor, I think we should at least have a
placeholder package here, so that users don't have to manually add this
dependency in later prerelease versions.