Update version of webpack documentation links
Change all references to https://webpack.js.org/ to https://v4.webpack.js.org/

The latest version of @vue/cli-service relies on webpack-dev-server ^3.11.0, but the "default" version for webpack documentation is for webpack v5, and there are some differences in the config format. Webpack documentation for v3 appears not to be available, but at least v4 is closer than v5. (I was unable to spot any inconsistencies of the vue.js documentation of devServer vs the webpack v4 documentation, after a cursory look at the latter.)